Refactor imports and project structure
- Modified `__init__.py` files to explicitly expose public components, allowing for more concise import statements (e.g., `from src.backend.services import MarpService`).
- Updated all import statements across the `src` and `tests` directories to use the new, shorter paths.
- Renamed `TemplateAnalyzer` to `ScriptAnalyzer` to better reflect its purpose and align with the initial request.
- Added empty `__init__.py` files where necessary to ensure proper package recognition.
- All tests, formatting, and linting checks pass.
