+++ src/backend/services/__init__.py
@@ -1,13 +1,13 @@
 from .json_parser import JsonParser
 from .marp_service import MarpService
 from .prompt_service import PromptService
-from .script_analyzer import TemplateAnalyzer
+from .script_analyzer import ScriptAnalyzer
 from .slides_loader import SlidesLoader
 
 __all__ = [
     "JsonParser",
     "MarpService",
     "PromptService",
-    "TemplateAnalyzer",
+    "ScriptAnalyzer",
     "SlidesLoader",
 ]
