+++ tests/backend/services/test_slides_loader.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from src.backend.services.slides_loader import SlidesLoader
+from src.backend.services import SlidesLoader
 
 
 class TestSlidesLoader:
